\name{kidney}
\alias{kidney}
\docType{data}
\title{
Kidney data.
}
\description{
This data set records age and kidney function of \eqn{N = 157} volunteers. Higher scores indicates better function.
}
\usage{data(kidney)}
\format{
  A data frame with 157 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{A list of patients' age.}
    \item{\code{z}}{A list of kidney scores.}
  }
}
\source{
http://statweb.stanford.edu/~ckirby/brad/LSI/datasets-and-programs/datasets.html
}
\references{
Efron, B. (2012). "Large-scale inference: empirical Bayes methods for estimation, testing, and prediction". Cambridge University Press.

Lemley, K. V., Lafayette, R. A., Derby, G., Blouch, K. L., Anderson, L., Efron, B., & Myers, B. D. (2007). "Prediction of early progression in recently diagnosed IgA nephropathy." Nephrology Dialysis Transplantation, 23(1), 213-222.
}
\keyword{datasets}