\name{LPRelevance-package}
\alias{LPRelevance-package}
\alias{LPRelevance}
\alias{LPcden}
\alias{eLP.poly}
\alias{eLP.univar}
\alias{LPregression}
\alias{Predict.LP.poly}
\alias{eLP.poly.predict}
\alias{LP.smooth}

\docType{package}
\title{
\packageTitle{LPRelevance}
}
\description{
How to individualize a global inference method? The goal of this package is to provide a systematic
recipe for converting classical global inference algorithms into customized ones. It provides
methods that perform individual level inferences by taking contextual covariates into
account. At the heart of our solution is the concept of "artificially-designed relevant samples",
called LASERs--which pave the way to construct an inference mechanism that is simultaneously
efficiently estimable and contextually relevant, thus works at both macroscopic (overall
simultaneous) and microscopic (individual-level) scale.
}

\author{
\packageAuthor{LPRelevance}

Maintainer: \packageMaintainer{LPRelevance}
}
\references{
Mukhopadhyay, S., and Wang, K (2021) "On The Problem of Relevance in Statistical Inference". <arXiv:2004.09588>
}

\keyword{ package }
