\name{funnel}
\alias{funnel}
\docType{data}
\title{
Simulation data set.
}
\description{
A simulated heterogeneous data set.
}
\usage{data("funnel")}
\format{
  A data frame with 3565 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{A list of covariate values.}
    \item{\code{z}}{A list of z-values.}
    \item{\code{tags}}{Binary vector of labels, 1 indicates a data point is a signal.}
  }
}
\references{
Mukhopadhyay, S., and Wang, K (2019) "On The Problem of Relevance in Statistical Inference". Technical Report. 
}

\keyword{datasets}
