\name{LPRelevance-package}
\alias{LPRelevance-package}
\alias{LPRelevance}
\alias{LPcden}
\alias{eLP.poly}
\alias{eLP.univar}
\alias{LP.smooth}
\docType{package}
\title{
\packageTitle{LPRelevance}
}
\description{
How to individualize a global inference model? The goal of this package is to provide a systematic 
recipe for converting a classical global inference algorithm into a customized one. It provides 
methods that perform individual level inferences by taking contextually relevant covariates into 
account. At the heart of our solution is the concept of "artificially-designed relevant samples", 
a.k.a. LASER. LASERs pave the way to construct an inference mechanism that is simultaneously 
efficiently estimable and contextually relevant, thus works at both macroscopic (overall 
simultaneous) and microscopic (individual-level) scale.
}

\author{
Mukhopadhyay, S. and Wang, K.

Maintainer: Kaijun Wang <kaijunwang.19@gmail.com>
}
\references{
Mukhopadhyay, S., and Wang, K (2019) "On The Problem of Relevance in Statistical Inference". Technical Report. 
}

\keyword{ package }